Impl logical addresses
Implemented mapping of logical addresses to the auto-increment ones to
create the data attribute if the IPAL messages in parse_packet().
The function now maintains a state of the current FMMUs of the slaves,
which store the mapping of logical addresses, in _fmmu_entities_map.
Implement processing of data of PDUs with configured addressing (FPWR,
FPRW) in parse_packet().
Add two new address types BROADCAST_ADDR and LOGICAL_ADDR.
Change the names of variables referring to configured addresses from
"phys_addr" to "config_addr".